chore: remove old unused files
