--- src/components/styles/FilterField.style.js
@@ -1,14 +1,0 @@
-import css from 'styled-jsx/css'
-import { colors } from '@dhis2/ui'
-
-export default css`
-    .container {
-        border: 1px solid ${colors.grey300};
-        background-color: ${colors.white};
-        height: 53px;
-        border-right: 0;
-        border-left: 0;
-        border-top: 0;
-        padding: 10px 5px 10px 10px;
-    }
-`
