+++ src/index.js
@@ -1,7 +1,7 @@
 import './locales'
 // Components
 
-export { default as DataDimension } from './components/DataDimension/DataDimensionNew'
+export { default as DataDimension } from './components/DataDimension/DataDimension'
 
 export { default as PeriodDimension } from './components/PeriodDimension/PeriodDimension'
 export { default as FixedPeriodSelect } from './components/PeriodDimension/FixedPeriodSelect'
